+++ relay-profiling/src/android.rs
@@ -1,3 +1,11 @@
+//! Android Format
+//!
+//! Relay is expecting a JSON object with some mandatory metadata and a `sampled_profile` key
+//! containing the raw Android profile.
+//!
+//! `android` has a specific binary representation of its profile and Relay is responsible to
+//! unpack it before it's forwarded down the line.
+//!
 use std::collections::{BTreeMap, HashMap};
 
 use android_trace_log::chrono::{DateTime, Utc};
@@ -8,7 +16,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::measurements::Measurement;
 use crate::native_debug_image::NativeDebugImage;
-use crate::sample::SampleProfile;
+use crate::sample::v1::SampleProfile;
 use crate::transaction_metadata::TransactionMetadata;
 use crate::utils::{deserialize_number_from_string, is_zero};
 use crate::{ProfileError, MAX_PROFILE_DURATION};
@@ -241,7 +249,7 @@ mod tests {
 
     #[test]
     fn test_roundtrip_android() {
-        let payload = include_bytes!("../tests/fixtures/profiles/android/roundtrip.json");
+        let payload = include_bytes!("../tests/fixtures/android/roundtrip.json");
         let profile = parse_profile(payload);
         assert!(profile.is_ok());
         let data = serde_json::to_vec(&profile.unwrap());
@@ -252,7 +260,7 @@ mod tests {
 
     #[test]
     fn test_roundtrip_react_native() {
-        let payload = include_bytes!("../tests/fixtures/profiles/android/roundtrip.rn.json");
+        let payload = include_bytes!("../tests/fixtures/android/roundtrip.rn.json");
         let profile = parse_profile(payload);
         assert!(profile.is_ok());
         let data = serde_json::to_vec(&profile.unwrap());
@@ -263,23 +271,21 @@ mod tests {
 
     #[test]
     fn test_no_transaction() {
-        let payload = include_bytes!("../tests/fixtures/profiles/android/no_transaction.json");
+        let payload = include_bytes!("../tests/fixtures/android/no_transaction.json");
         let data = parse_android_profile(payload, BTreeMap::new(), BTreeMap::new());
         assert!(data.is_err());
     }
 
     #[test]
     fn test_remove_invalid_events() {
-        let payload =
-            include_bytes!("../tests/fixtures/profiles/android/remove_invalid_events.json");
+        let payload = include_bytes!("../tests/fixtures/android/remove_invalid_events.json");
         let data = parse_android_profile(payload, BTreeMap::new(), BTreeMap::new());
         assert!(data.is_err());
     }
 
     #[test]
     fn test_transactions_to_top_level() {
-        let payload =
-            include_bytes!("../tests/fixtures/profiles/android/multiple_transactions.json");
+        let payload = include_bytes!("../tests/fixtures/android/multiple_transactions.json");
 
         let profile = match parse_profile(payload) {
             Err(err) => panic!("cannot parse profile: {err:?}"),
@@ -308,7 +314,7 @@ mod tests {
             ),
         ]);
 
-        let payload = include_bytes!("../tests/fixtures/profiles/android/valid.json");
+        let payload = include_bytes!("../tests/fixtures/android/valid.json");
         let profile_json = parse_android_profile(payload, transaction_metadata, BTreeMap::new());
         assert!(profile_json.is_ok());
 
@@ -330,7 +336,7 @@ mod tests {
 
     #[test]
     fn test_timestamp() {
-        let payload = include_bytes!("../tests/fixtures/profiles/android/roundtrip.json");
+        let payload = include_bytes!("../tests/fixtures/android/roundtrip.json");
         let profile = parse_profile(payload);
 
         assert!(profile.is_ok());