+++ tests/integration/test_metrics.py
@@ -858,7 +858,7 @@ def test_transaction_metrics(
 
         return
 
-    metrics = metrics_by_name(metrics_consumer, count=10, timeout=6)
+    metrics = metrics_by_name(metrics_consumer, count=11, timeout=6)
 
     common = {
         "timestamp": int(timestamp.timestamp()),
@@ -1430,10 +1430,14 @@ def test_span_metrics(
         for metric, headers in metrics
         if metric["name"].startswith("spans", 2)
     ]
-    assert len(span_metrics) == 6
+    assert len(span_metrics) == 8
     for metric, headers in span_metrics:
         assert headers == [("namespace", b"spans")]
-        if metric["name"] in ("c:spans/count_per_op@none", "c:spans/usage@none"):
+        if metric["name"] in (
+            "c:spans/count_per_op@none",
+            "c:spans/usage@none",
+            "d:spans/duration@millisecond",
+        ):
             continue
 
         # Ignore transaction spans